add Pod

+++ Classes/Controllers/LXUserPageViewController.m
@@ -1001,7 +1001,7 @@ typedef enum {
     
     [operation setCompletionBlockWithSuccess: successUpload failure: failUpload];
     
-    [operation setUploadProgressBlock:^(NSUInteger bytesWritten, NSInteger totalBytesWritten, NSInteger totalBytesExpectedToWrite) {
+    [operation setUploadProgressBlock:^(NSUInteger bytesWritten, long long totalBytesWritten, long long totalBytesExpectedToWrite) {
         progessHUD.progress = (float)totalBytesWritten/(float)totalBytesExpectedToWrite;
     }];
